Moving logic to custom hook
